docs: add quick start instructions
