+++ .env.example
@@ -3,9 +3,3 @@ TCP_PORT=8899
 
 API_HOST=0.0.0.0
 API_PORT=8000
-
-MYSQL_HOST=127.0.0.1
-MYSQL_PORT=3306
-MYSQL_USER=root
-MYSQL_PASSWORD=123456
-MYSQL_DB=alarm
