--- config.py
@@ -1,16 +1,0 @@
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-TCP_HOST = os.getenv("TCP_HOST", "0.0.0.0")
-TCP_PORT = int(os.getenv("TCP_PORT", "8899"))
-
-API_HOST = os.getenv("API_HOST", "0.0.0.0")
-API_PORT = int(os.getenv("API_PORT", "8000"))
-
-MYSQL_HOST = os.getenv("MYSQL_HOST", "127.0.0.1")
-MYSQL_PORT = int(os.getenv("MYSQL_PORT", "3306"))
-MYSQL_USER = os.getenv("MYSQL_USER", "root")
-MYSQL_PASSWORD = os.getenv("MYSQL_PASSWORD", "123456")
-MYSQL_DB = os.getenv("MYSQL_DB", "alarm")
