+++ requirements.txt
@@ -1,4 +1,3 @@
 fastapi
 uvicorn
-pymysql
 python-dotenv
